+++ composeApp/src/commonMain/kotlin/jp/kyamlab/todolist/di/AppModule.kt
@@ -1,13 +1,14 @@
 package jp.kyamlab.todolist.di
 
 import jp.kyamlab.todolist.data.ToDoRepository
+import jp.kyamlab.todolist.data.ToDoRepositoryImpl
 import jp.kyamlab.todolist.ui.archive.ArchiveViewModel
 import jp.kyamlab.todolist.ui.home.HomeViewModel
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 
 val appModule = module {
-    single { ToDoRepository() }
+    single<ToDoRepository> { ToDoRepositoryImpl() }
     viewModel { HomeViewModel(get()) }
     viewModel { ArchiveViewModel(get()) }
 }
